Scaled the data, found a pretty good k, created the confusion matrix
For k nearest neighbors, I'm pretty sure we can only use numeric predictors. So I chose to use all of the numeric predictors except for slope_peak. Then I scaled all of them.

As for finding the optimal k, it seems that every time I run the code the optimal k seems to change a bit. I also tried changing the range of values of k up to 50 and got a lot of different answers. For now I guess we can settle with k = 12.

I'm also not super sure how to do the data visualization. Since we have 5 predictors, we can't exactly graph it on a 2d plane. Maybe we just pick two variables to graph?