Summarized data and visualized training data
I realized that the proposal doesn't actually require any actual analysis of the data yet. So I made the summary using the number of observations of yes and no presence, as well as the means of each predictor. I looked for na's, but it seems this dataset uses ?, and I'm not too sure how to count the number of appearances of ?. Either way, they only appeared in columns 12 and 13 anyways.

For the visualization, I just made distribution histograms of each predictor. I still need to make them look nicer with labels and whatnot.

I also tried doing the analysis with pres_val instead of presence. The model wasn't able to predict anything with pres_val=4 so I think it's better to stick with presence since it only uses "yes" and "no".